requirements.txt
requests==2.32.3
python-chess==1.999

fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2
python-multipart>=0.0.9

httpx>=0.27.0
